# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "api autentia manager",
        "title": "Autentia Manager Documentation",
        "contact": {},
        "version": "1"
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "login",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.loginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inició sesión correctamente"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/activate": {
            "post": {
                "description": "activate user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "activate",
                "parameters": [
                    {
                        "description": "activate",
                        "name": "activate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.userValidateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.responseValidation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
        "/users/confirm": {
            "post": {
                "description": "confirm user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Confirm",
                "parameters": [
                    {
                        "description": "confirm",
                        "name": "recovery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.validateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmación de usuario exitosa"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/guest": {
            "post": {
                "description": "login guest",
                "tags": [
                    "login"
                ],
                "summary": "login guest",
                "responses": {
                    "200": {
                        "description": "Inició sesión correctamente"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/recovery": {
            "post": {
                "description": "recovery pass",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Recovery",
                "parameters": [
                    {
                        "description": "recovery",
                        "name": "recovery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.recoveryParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Se enviará correo para recuperar su contraseña a la dirección example@mail.com"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/validate": {
            "post": {
                "description": "validate user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Validate",
                "parameters": [
                    {
                        "description": "validate",
                        "name": "validate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.userValidateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.responseValidation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{nickName}/reconfirmation": {
            "post": {
                "description": "user reconfirmation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Reconfirmation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nickName",
                        "name": "nickName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.Country": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "data.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "country": {
                    "$ref": "#/definitions/data.Country"
                },
                "dni": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "status_user": {
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
        "routes.MessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                }
            }
        },
...
        "routes.loginParams": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "required": [
                        "nickname",
                        "password"
                    ],
                    "properties": {
                        "nickname": {
                            "type": "string"
                        },
                        "password": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "routes.recoveryParams": {
            "type": "object",
            "required": [
                "email",
                "nickname"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "routes.response": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/data.User"
                }
            }
        },
...
        "routes.responseValidation": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "routes.userValidateParams": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "routes.validateParams": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "recovery": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "barerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}